fix: rename file and move ouTitle fn to modules
