+++ src/visualizations/config/adapters/dhis_highcharts/getFilterText.js
@@ -1,9 +1,9 @@
 import isArray from 'd2-utilizr/lib/isArray'
-import i18n from '@dhis2/d2-i18n'
 import { ouIdHelper } from '../../../../modules/ouIdHelper'
 import { dimensionIs } from '../../../../modules/layout/dimensionIs'
 import { DIMENSION_ID_ORGUNIT } from '../../../../modules/fixedDimensions'
 import { dimensionGetItems } from '../../../../modules/layout/dimensionGetItems'
+import { getOuFilterText } from '../../../../modules/ouFilterText'
 
 export default function(filters, metaData) {
     if (isArray(filters)) {
@@ -22,11 +22,11 @@ export default function(filters, metaData) {
 
             if (hasOuLevel || hasOuGroup) {
                 if (hasOuGroup) {
-                    titleParts.push(getOuTitle(items, metaData, false))
+                    titleParts.push(getOuFilterText(items, metaData, false))
                 }
 
                 if (hasOuLevel) {
-                    titleParts.push(getOuTitle(items, metaData, true))
+                    titleParts.push(getOuFilterText(items, metaData, true))
                 }
             } else {
                 const filterItems = metaData.dimensions[filter.dimension]
@@ -65,43 +65,4 @@ export default function(filters, metaData) {
     }
 
     return null
-}
-
-const getOuTitle = (items, metaData, isLevel) => {
-    const getNameFromMetadata = id =>
-        metaData.items[id] ? metaData.items[id].name : id
-
-    const dynamicOuItems = items.filter(item =>
-        isLevel
-            ? ouIdHelper.hasLevelPrefix(item.id)
-            : ouIdHelper.hasGroupPrefix(item.id)
-    )
-    const lastItem = dynamicOuItems.length > 1 ? dynamicOuItems.pop() : null
-    const dynamicOuNames = dynamicOuItems
-        .map(item => getNameFromMetadata(ouIdHelper.removePrefix(item.id)))
-        .join(', ')
-
-    let allDynamicOuNames
-
-    if (lastItem) {
-        const lastOuName = getNameFromMetadata(
-            ouIdHelper.removePrefix(lastItem.id)
-        )
-        allDynamicOuNames = `${dynamicOuNames} ${i18n.t('and')} ${lastOuName}`
-    } else {
-        allDynamicOuNames = dynamicOuNames
-    }
-
-    const staticOuNames = items
-        .filter(
-            item =>
-                !ouIdHelper.hasGroupPrefix(item.id) &&
-                !ouIdHelper.hasLevelPrefix(item.id)
-        )
-        .map(item => getNameFromMetadata(item.id))
-        .join(', ')
-
-    const joiner = isLevel ? i18n.t('levels in') : i18n.t('groups in')
-
-    return `${allDynamicOuNames} ${joiner} ${staticOuNames}`
 }
