+++ src/visualizations/config/adapters/dhis_highcharts/title/yearOverYear.js
@@ -1,14 +1,14 @@
-import getFilterTitle from '../getFilterTitle'
+import getFilterText from '../getFilterText'
 
 export default function(layout, metaData, dashboard) {
     const titleFragments = []
 
     if (layout.columns && !dashboard) {
-        titleFragments.push(getFilterTitle(layout.columns, metaData))
+        titleFragments.push(getFilterText(layout.columns, metaData))
     }
 
     if (layout.filters && !dashboard) {
-        titleFragments.push(getFilterTitle(layout.filters, metaData))
+        titleFragments.push(getFilterText(layout.filters, metaData))
     }
 
     return titleFragments.length ? titleFragments.join(' - ') : null
